+++ cmd/system.go
@@ -3,17 +3,22 @@ package cmd
 import (
 	"errors"
 	"fmt"
+	"io"
 	"log"
 	"net"
 	"os"
+	"path"
 	"reflect"
+	"regexp"
 	"sort"
 	"strconv"
 	"strings"
 
 	"github.com/Jeffail/gabs/v2"
+	"github.com/kevinburke/ssh_config"
 	"github.com/level27/l27-go"
 	"github.com/level27/lvl/utils"
+	homedir "github.com/mitchellh/go-homedir"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"golang.org/x/sync/errgroup"
@@ -247,6 +252,9 @@ func init() {
 	// SYSTEM SCP
 	systemCmd.AddCommand(systemScpCommand)
 
+	// SYSTEM SSHCONFIG
+	systemCmd.AddCommand(systemSshConfigCmd)
+
 	//------------------------------------- NETWORKS -------------------------------------
 	// #region NETWORKS
 
@@ -456,9 +464,9 @@ func resolveSystemVolume(systemID l27.IntID, arg string) (l27.IntID, error) {
 	return ip.ID, nil
 }
 
-//------------------------------------------------- SYSTEM TOPLEVEL (GET / DESCRIBE CREATE) ----------------------------------
+// ------------------------------------------------- SYSTEM TOPLEVEL (GET / DESCRIBE CREATE) ----------------------------------
 // #region SYSTEM TOPLEVEL (GET / DESCRIBE / CREATE)
-//----------------------------------------- GET ---------------------------------------
+// ----------------------------------------- GET ---------------------------------------
 var systemGetCmd = &cobra.Command{
 	Use:   "get",
 	Short: "get a list of all curent systems",
@@ -494,7 +502,7 @@ func getSystems(ids []l27.IntID) ([]l27.System, error) {
 	return systems, nil
 }
 
-//----------------------------------------- DESCRIBE ---------------------------------------
+// ----------------------------------------- DESCRIBE ---------------------------------------
 var systemDescribeHideJobs = false
 
 var systemDescribeCmd = &cobra.Command{
@@ -563,7 +571,7 @@ var systemDescribeCmd = &cobra.Command{
 	},
 }
 
-//----------------------------------------- CREATE ---------------------------------------
+// ----------------------------------------- CREATE ---------------------------------------
 // vars needed to save flag data.
 var systemCreateName, systemCreateFqdn, systemCreateRemarks string
 var systemCreateDisk, systemCreateCpu, systemCreateMemory int32
@@ -695,7 +703,7 @@ var systemCreateCmd = &cobra.Command{
 
 // #endregion
 
-//------------------------------------------------- SYSTEM SPECIFIC (UPDATE / FORCE DELETE ) ----------------------------------
+// ------------------------------------------------- SYSTEM SPECIFIC (UPDATE / FORCE DELETE ) ----------------------------------
 // #region SYSTEM SPECIFIC (UPDATE / FORCE DELETE)
 var systemUpdateCmd = &cobra.Command{
 	Use:   "update",
@@ -1079,7 +1087,7 @@ var systemCheckUpdateCmd = &cobra.Command{
 
 // #endregion
 
-//------------------------------------------------- MONITORING ON SPECIFIC SYSTEM ----------------------------------------------
+// ------------------------------------------------- MONITORING ON SPECIFIC SYSTEM ----------------------------------------------
 // ---- MONITORING COMMAND
 var systemMonitoringCmd = &cobra.Command{
 	Use:     "monitoring",
@@ -1207,7 +1215,7 @@ func runAction(action string, args []string) error {
 
 // #endregion
 
-//------------------------------------------------- SYSTEM/COOKBOOKS TOPLEVEL (GET error / CREATE) ----------------------------------
+// ------------------------------------------------- SYSTEM/COOKBOOKS TOPLEVEL (GET error / CREATE) ----------------------------------
 // ---------------- MAIN return COMMAND (cookbooks)
 var systemCookbookCmd = &cobra.Command{
 	Use:   "cookbooks",
@@ -1419,10 +1427,10 @@ var systemCookbookDeleteCmd = &cobra.Command{
 
 // ---------------- UPDATE
 var systemCookbookUpdateCmd = &cobra.Command{
-	Use:   "update [systemID] [cookbookID]",
-	Short: "update existing cookbook from a system",
+	Use:     "update [systemID] [cookbookID]",
+	Short:   "update existing cookbook from a system",
 	Example: "lvl system cookbooks update [systemID] [cookbookID] {-p}.\nSINGLE PARAMETER:		-p waf=true  \nMULTIPLE PARAMETERS:		-p waf=true -p timeout=200  \nMULTIPLE VALUES:		-p versions=''7, 5.4'' OR -p versions=7,5.4 (seperated by comma)",
-	Args: cobra.ExactArgs(2),
+	Args:    cobra.ExactArgs(2),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// check for valid system id
 		systemID, err := resolveSystem(args[0])
@@ -1599,7 +1607,7 @@ func CheckCBValueForParameter(value string, options l27.CookbookParameterOptionV
 
 // #endregion
 
-//------------------------------------------------- SYSTEMS/GROUPS (GET / ADD  / DELETE)-------------------------------------------------
+// ------------------------------------------------- SYSTEMS/GROUPS (GET / ADD  / DELETE)-------------------------------------------------
 // ---------------- MAIN COMMAND (groups)
 var SystemSystemgroupsCmd = &cobra.Command{
 	Use:   "groups",
@@ -1801,8 +1809,15 @@ var systemSshKeysRemoveCmd = &cobra.Command{
 
 // SYSTEM SSH
 var systemSshCmd = &cobra.Command{
-	Use:   "ssh [system] [ssh args]",
+	Use:   "ssh <system> [flags] [--] [ssh args]",
 	Short: "Connect to a system via SSH, automatically adding SSH keys to the system if necessary",
+	Long: `Connect to a system via SSH, automatically adding SSH keys to the system if necessary.
+The command will automatically add your favorite SSH key to the system if necessary. You'll need to use 'lvl sshkey favorite' to configure it the first time.
+The command figures out a valid IP address to connect to and passes it through to the ssh command.
+Arguments passed after the system ID/name are passed to ssh literally. Note that for any flags starting with "-", you'll want to put "--" before them so lvl does not try to interpret them as flags itself.`,
+	Example: `lvl system ssh my-awesome-server
+lvl system ssh my-awesome-server ls "~"
+lvl system ssh my-awesome-server -- ls -l "~"`,
 
 	Args: cobra.MinimumNArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
@@ -1894,6 +1909,8 @@ func waitAddSshKey(systemID l27.IntID, sshKeyID l27.IntID) error {
 }
 
 // Resolve the hostname to SSH into a system.
+// If the FQDN properly resolves, we use that.
+// Otherwise we try the IP addresses in the system's networks.
 func sshResolveHost(systemID l27.IntID) (string, error) {
 	system, err := Level27Client.SystemGetSingle(systemID)
 	if err != nil {
@@ -2040,6 +2057,236 @@ var systemScpCommand = &cobra.Command{
 
 		return tailExecProcess("scp", scpArgs)
 	},
+}
+
+var systemSshConfigCmd = &cobra.Command{
+	Use:   "sshconfig",
+	Short: "Add system's name to your user SSH config for easy access",
+	Long: `Add system's name to your user SSH config for easy access
+This will add a Host entry to your SSH config, so afterwards you can use commands outside lvl to access the system by name.
+For example: rsync foo.txt my-awesome-system:~/
+
+The new host names are written into a separate ~/.ssh/lvl config file, which gets added to your ~/.ssh/config via an Include directive.`,
+
+	Args: cobra.ExactArgs(1),
+	RunE: func(cmd *cobra.Command, args []string) error {
+		systemID, err := resolveSystem(args[0])
+		if err != nil {
+			return err
+		}
+
+		address, err := sshResolveHost(systemID)
+		if err != nil {
+			return err
+		}
+
+		system, err := Level27Client.SystemGetSingle(systemID)
+		if err != nil {
+			return err
+		}
+
+		f, configPath, err := ensureSshConfig()
+		if err != nil {
+			return err
+		}
+
+		var cfg *ssh_config.Config
+		(func() {
+			defer f.Close()
+
+			cfg, err = ssh_config.Decode(f)
+		})()
+
+		if err != nil {
+			return fmt.Errorf("unable to parse SSH config: %s", err.Error())
+		}
+
+		host := findSshConfigHost(cfg, system.Name)
+		if host == nil {
+			pattern, err := ssh_config.NewPattern(system.Name)
+			if err != nil {
+				return fmt.Errorf("invalid SSH config host value: '%s'", system.Name)
+			}
+
+			// Host not in config file yet, add a new one!
+			host = &ssh_config.Host{
+				Patterns: []*ssh_config.Pattern{pattern},
+			}
+
+			cfg.Hosts = append(cfg.Hosts, host)
+		}
+
+		host.EOLComment = fmt.Sprintf(" system ID: %d", systemID)
+
+		setSshConfigHostNode(host, "HostName", address)
+		setSshConfigHostNode(host, "User", "root")
+
+		// Write new config back to file.
+		err = os.WriteFile(configPath, []byte(cfg.String()), 0o644)
+		if err != nil {
+			return fmt.Errorf("failed to write %s: %s", configPath, err.Error())
+		}
+
+		err = ensureSshIncludeWritten()
+		if err != nil {
+			return err
+		}
+
+		outputFormatTemplate(system, "templates/entities/system/sshConfigConfirm.tmpl")
+
+		return nil
+	},
+}
+
+func findSshConfigHost(cfg *ssh_config.Config, name string) *ssh_config.Host {
+	for _, host := range cfg.Hosts {
+		for _, pattern := range host.Patterns {
+			if pattern.String() == name {
+				return host
+			}
+		}
+	}
+
+	return nil
+}
+
+func setSshConfigHostNode(host *ssh_config.Host, key string, value string) {
+	for _, node := range host.Nodes {
+		kv, ok := node.(*ssh_config.KV)
+		if !ok {
+			continue
+		}
+
+		if kv.Key == key {
+			kv.Value = value
+			return
+		}
+	}
+
+	// Didn't find the key in the config yet, add a new one.
+	host.Nodes = append(host.Nodes, &ssh_config.KV{
+		Key:   key,
+		Value: value,
+	})
+}
+
+func ensureSshConfig() (*os.File, string, error) {
+	return ensureSshDirFileExists(getSshConfigFileName())
+}
+
+// Ensure there is an include directive to lvl's SSH config file in the user's ~/.ssh/config.
+func ensureSshIncludeWritten() error {
+	f, fullPath, err := ensureSshDirFileExists("config")
+	if err != nil {
+		return fmt.Errorf("error opening SSH config file: %s", err.Error())
+	}
+
+	var configData []byte
+	(func() {
+		defer f.Close()
+
+		configData, err = io.ReadAll(f)
+	})()
+	if err != nil {
+		return fmt.Errorf("error reading SSH config file: %s", err.Error())
+	}
+
+	configName := getSshConfigFileName()
+	regex := fmt.Sprintf(`Include\s+%s\b`, regexp.QuoteMeta(configName))
+	matched, err := regexp.Match(regex, configData)
+	if err != nil {
+		return fmt.Errorf(
+			"error checking for existing include directive in SSH config file: %s",
+			err.Error())
+	}
+
+	if matched {
+		// Already has the include directive, don't need to do anything!
+		return nil
+	}
+
+	fmt.Fprintf(os.Stderr, "Adding Include for lvl config file to SSH config\n")
+
+	if len(configData) != 0 {
+		// Copy SSH config file as a backup in case we screw up.
+		err = (func() error {
+			backupPath := fullPath + ".bak"
+			fmt.Fprintf(os.Stderr, "Copying %s to %s as backup", fullPath, backupPath)
+
+			backup, err := os.Create(backupPath)
+			if err != nil {
+				return fmt.Errorf("failed to open %s: %s", backupPath, err.Error())
+			}
+
+			backup.Write(configData)
+			return nil
+		})()
+
+		if err != nil {
+			return err
+		}
+	}
+
+	f, err = os.OpenFile(fullPath, os.O_RDWR, 0)
+	if err != nil {
+		return fmt.Errorf("failed to open %s for writing: %s", fullPath, err.Error())
+	}
+
+	defer f.Close()
+
+	includeBytes := []byte(fmt.Sprintf("Include %s\n\n", configName))
+	_, err = f.Write(includeBytes)
+
+	if err != nil {
+		return fmt.Errorf("error while writing to %s: %s", fullPath, err.Error())
+	}
+
+	return nil
+}
+
+// Ensure ~/.ssh/<name> exists, or create it (and the parent directory) if necessary.
+func ensureSshDirFileExists(name string) (*os.File, string, error) {
+	home, err := homedir.Dir()
+	if err != nil {
+		return nil, "", fmt.Errorf("unable to determine home directory: %s", err.Error())
+	}
+
+	path_ssh := path.Join(home, ".ssh")
+	ssh_config := path.Join(path_ssh, name)
+	f, err := os.Open(ssh_config)
+	if os.IsNotExist(err) {
+		// Automatically make an empty ~/.ssh/<config file>.
+		_, err = os.Stat(path_ssh)
+		if os.IsNotExist(err) {
+			// Have to make ~/.ssh/ too...
+			err = os.Mkdir(path_ssh, 0o700)
+			if err != nil {
+				return nil, "", fmt.Errorf("failed to create %s: %s", path_ssh, err.Error())
+			}
+		}
+
+		// Make ~/.ssh/config.
+		f, err = os.Create(ssh_config)
+		if err != nil {
+			return nil, "", fmt.Errorf("failed to create %s: %s", ssh_config, err.Error())
+		}
+	} else if err != nil {
+		return nil, "", fmt.Errorf("failed to open %s: %s", ssh_config, err.Error())
+	}
+
+	return f, ssh_config, nil
+}
+
+// Get the file name to write system SSH config entries to.
+// This is not ~/.ssh/config itself, but we write an include directive into ~/.ssh/config for it.
+func getSshConfigFileName() string {
+	name := viper.GetString("ssh_config_name")
+	if name != "" {
+		return name
+	}
+
+	// No value given, return default.
+	return "lvl"
 }
 
 // NETWORKS